Quite OK ATT flight

+++ sw/airborne/modules/sensors/cameras/jevois_mavlink.h
@@ -61,7 +61,7 @@ static inline void comm_send_ch(mavlink_channel_t chan __attribute__((unused)), 
 
 struct vision_relative_position_struct {
   int received;
-  int cnt;
+  uint16_t cnt;
   float x;
   float y;
   float z;
